requirements.txt
# Core dependencies
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.0

# Plaid API
plaid-python==15.0.0

# ML & Data Processing
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.3
joblib==1.3.2

# Dashboard & Visualization
streamlit==1.29.0
plotly==5.18.0
altair==5.2.0

# Utilities
cryptography==41.0.7
apscheduler==3.10.4
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==23.12.0
pylint==3.0.3
mypy==1.7.1